blogApp/requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.1.13
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
mongoengine==0.27.0
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
pymongo==3.12.3
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.2.4
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
